feat : [FE] PlaybackTools 기능관련 부분 이동
- PlaybackToolClass로 옮겨 controller에서 사용할 수 있도록 함.
- 정지와 반복을 제외한 모든 키는 단축키로 사용할 수 있음.
- mute, solo에 대한 정보를 지역적으로 가지고 있음.
